bump tolerance on unit test.
This is an old test from way back in the porting process. I haven't
found any reference to this test in Karney's test suite, so I'm assuming
it's more of a regression test from the original porter than it is a
specification of the desired behavior, and thus feel ok about bumping
the tolerance.

Note this seemed to start failing when switching from the rust:1.66
container to using "latest rust" on "ubuntu-latest" container.