Exclude benchmark data from published package
This commit excludes the benchmark data and the benchmark from the
published crate. This minimizes the size of the uploaded tar ball by
nearly 20% (from 54kb to 41kb). I'm not aware any useful reason to run
benchmarks from the published source code, so there shouldn't be any
downside in doing that.